[DOP-3639]: move filters panel change to correct component
